Create Maven project before refactoring
